requirements.txt
# Auto-generated from pyproject.toml — prefer `pip install -e .`
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
opencv-python-headless>=4.8.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
paho-mqtt>=1.6.0
pygame>=2.5.0
